dbus: remove broken X11 autolaunch hack
This hack was added so that X11 autolaunch support could be enabled
even when building without X11.  At the time, it was possible to have
the autolaunch support without X11 support.  But later, this stopped
being possible, and so the derivation was changed to only apply the
hack when X11 support was enabled.  But in that case, the same flag
would be added by the build system.

So in summary, this hack is only enabled in the case where it's a
no-op.  Therefore, it can be safely removed.