linuxPackages.rtl8821au: 2021-05-18 -> 2021-11-05
Fixes build with Linux 5.15.

The repo name has changed.  The old repo has changed to just a README
that tells you which versioned repository you should look at.  It's a
bit strange.  I've left the old repo as the homepage, because that's
the place to look while updating to see if there's a new code repo to
use.