nixos/modules/services/monitoring/zabbix-agent.nix: added UserParameters option
I wanted to add multiple userParameters, but the services.zabbixAgent.settings option
only allows a single Key value pair.
This cannot be used in conjunction with an "Include" directive already present, as the
include will be used for a generated configuration file.
People that already use an include file will not be affected.