chromium: Install vk_swiftshader_icd.json
This should be required for Vulkan support via SwiftShader but I wasn't
able to test/confirm it yet (according to strace Chromium never tried to
load that file). The relevant flags for using Chromium with SwiftShader
are documented at [0] and SwiftShader's ANGLE documentation [1] mentions
the relevant files. See also [2], [3], and [4] for more details.

[0]: https://source.chromium.org/chromium/chromium/src/+/master:docs/gpu/swiftshader.md;drc=046f987e020baba45ffb3061b3ee3d960d6ce981
[1]: https://github.com/google/swiftshader/blob/241349107851000f6a48c3d5769ed617785c8962/docs/ANGLE.md
[2]: https://github.com/google/swiftshader/blob/241349107851000f6a48c3d5769ed617785c8962/README.md
[3]: https://github.com/KhronosGroup/Vulkan-Loader/blob/ca9efe3f1ccd98a5fe05da71bf65087bcaa82f35/docs/LoaderInterfaceArchitecture.md
[4]: https://bugs.archlinux.org/task/72652