hercules-ci-{agent,cnix-expr}: Force use of propagated boehmgc
This was already in effect, but only by coincidence. A direct
boehmgc could presumably be added implicitly in the generated
code by fixing the pkg-config to nixpkgs mapping.

This commit should counteract that, or, in case of a rename,
cause an evaluation error so that it can be fixed, instead of
building and crashing at runtime.