qemu: 6.1.0 -> 6.2.0
QEMU now supports automatically detecting which audio drivers to
build, so we don't need to pass --audio-drv-list anymore.

pkgsMusl.qemu* doesn't build at the moment because of dependency
issues, but with a lot of hacks I was able to do enough of a build to
be confident the changes to patches etc. are correct for Musl.