Support XDG-conformant locations for the nix profile
Ref: https://github.com/NixOS/nix/pull/5588

This searches in $XDG_DATA_HOME/nix/profile everywhere where
previously we only looked in ~/.nix-profile. This shouldn't break
any compatibility for now. It is highly recommended to use nixpkgs
with these changes if you want to use Nix with the XDG basedir PR, but
not doing so shouldn't result in any dramatic breakage -- in fact,
NixOS is perfectly usable without any user profile.