retroarch: 1.9.2 -> 1.9.13.2
The issue of non-working cores on newer versions of RetroArch was caused
by the missing core metadata that is available on
libretro/libretro-super repo. This also allows RetroArch to works
properly, for example there is no need to load a core before loading a
content: RetroArch knows each emulator to load depending on the
available emulators and the file extension.

However, there were one issue (already presented on master) that if
the config file was created only on the first run RetroArch. Subsequent
runs could have out-of-date paths. Not a huge issue since if the path is
invalid RetroArch will go back to its defaults, but the inclusion of
metadata would make this issue worse since anyone that ever run
RetroArch once will have an unworking RetroArch, unless they fixed their
config to point to `libretro_info_path` to the correct path.

This commit also fixes it by ignoring the `libretro_directory` and
`libretro_info_path` configuration, and just hard-coding it to the path
on the `/nix/store`.