lrzip: enable asm on x86 and fix build on darwin
Enable compilation of native assembly code on x86 systems, instead of
disabling it for all systems. On darwin ensure the format is set to
macho64.

The ASM/x86 directory is compiled whether "--enable-asm" is configured
or not, but it creates an empty archive, which fails on darwin, so
ensure it is not compiled on darwin to fix the build.