Revert "sd-image-aarch64: Add console command line for Rock[Pro]64"
Due to complications with how consoles currently work in NixOS, this may
not work as expected in all cases. For now, we'll leave this change out.

This means that the generic AArch64 image needs to be modified to get
console output on RockPro64 and Rock64:

Mount the root partition (the larger one) on a PC, edit
boot/extlinux/extlinux.conf and replace
  console=ttyS0,115200n8
with
  console=ttyS2,1500000n8

It is probably also needed in the NixOS system configuration under
`boot.kernelParams`.

See discussion at https://github.com/NixOS/nixpkgs/pull/146725#discussion_r753626276