nixos/podman: Fix AppArmor support
Podman generates and loads its AppArmor profile at runtime, for this is
needs `apparmor_parser` to be in $PATH.

Podman detects at runtime if /etc/apparmor.d/tunables/global is present
and, if so, `include`s it in the runtime-generated profile. If it’s not
found then it does not, but then the profile fails to load successfully
because `…/abstractions/base` *is* found and that implicitly depends on
`…/tunables/global`.

Prevent future breakage with a dedicated test for Podman AppArmor
confinement with the default profile.