haskell-modules/lib: make postFixup more resilient
1. Putting the string on one line (instead of using a multiline
string) means it is not ended with \n, which means if somebody wants
to append more lines, they have to start it with \n or it will get
concatenated to the `postFixup` here, so add \n.

2. Previously, ignoring `drv.postFixup` would throw away any
`postFixup`s set in `drv`, which is not something we want.

The extra empty line is a cute trick to prevent concatenating the code
with any previous `postFixup` that doesn’t end in \n.