release-cross.nix: add jobs for missing lib.systems.examples
Adding something to lib.systems.examples means that it'll be exposed via
pkgsCross to users quite prominently. As a result we should try to
provide binary cache for at least the toolchains on a best effort
basis.

This commit syncs release-cross.nix with (new?) cross targets added to
lib.systems.examples in the meantime. We for example didn't have a
prebuilt toolchain for riscv64.