tests/sd-confinement: Add test case for UMask 0077
This is just in preparation to fixing #147599, because when using
confinement with a non-root user and a restrictive UMask, the leading
directories for all of the bind mounts we need in order to execute the
program in question will get that UMask as well.

In that issue, I also suggested a workaround, which is to set the
following option:

  TemporaryFileSystem = [ "/" "/nix" "/nix/store" ];

While this works, @martined has already opened an upstream issue at
https://github.com/systemd/systemd/issues/21548 so before we jump to
introduce such ugly things as the workaround above, let's first see what
upstream has to say.

Nevertheless, this test case is there so that once we have a definitive
answer, we already have a regression test.

Signed-off-by: aszlig <aszlig@nix.build>