tests/sd-confinement: Switch to description attr
This is only a code style change, since with the port to Python in
1320f23a6b5517e73bb886ffb23bec5b037914bf, the test became less readable
because the description attribute got inlined in the individual
testScript snippet in a series of "with subtest(...):".

While I do welcome the port to Python, I think this also made the code
harder to follow because every testScript now contains additional
boilerplate.

The reason why the descriptions got inlined was probably because at the
time we were using "Black" (a Python code formatter), which would barf
if a description would exceed a certain amount of characters.

However, since Black is no longer used and we switched to a linter a
long time ago (c362a28fcf2621fd3b6d6a96c821e9af9d123e21), so we can now
re-introduce the description attribute.

Signed-off-by: aszlig <aszlig@nix.build>