nixos/github-runner: use `!` for `ExecStartPre`
Use `!` instead of `+` to move and adapt the permissions of credentials
with `ExecStartPre=` scripts. The `!` prefix grants less privileges than
`+`. Also, `!` explicitly makes sure that the dynamic user/group pair
(`DynamicUser=`) is allocated before running the command. This is
necessary as we change the owner of files to the dynamically allocated
user.