nixos/github-runner: remove token before service start
An `ExecStartPre=` script copies the configured token to the
github-runner runtime directory to detect a configuration change to
trigger a new runner registration. After a runner registration, the
token is cleaned up. If, however, the configuration is unchanged, the
closing `ExecStartPre=` script did not delete the token in the runtime
directory. This commit causes the script to always delete the token
making it inaccessible to the main service process.