python3Packages.mahotas: fix freeimage support
use a much stronger binding to our specific freeimage that
works reliably on linux. previously it didn't and the tests
covering freeimage support were just being skipped as they
assumed it to be disabled.

once the binding works it reveals slight breakage in the tests
themselves, fixed with an upstream patch. this breakage was already
revealing itsef on darwin as the freeimage binding was "working"
there.