util-linux: Hack around pam's current lack of pkg-config
With the addition of audit to pam's buildInputs, it now has a transitive
dependency on it. This causes packages that attempt to statically link
pam to fail with missing audit symbol errors. Support for pkg-config is
in the next version of pam, but util-linux will also need to be updated
upstream to actually use pkg-config to find pam. Until then, this gets
it compiling statically again.