julia: switch from python2 to python3
Julia does not actually require python2; it just requires any version of python
[that is equal to or greater than 2.7](https://github.com/JuliaLang/julia/blob/master/doc/src/devdocs/build/build.md#required-build-tools-and-external-libraries).

This changes the python dependency from python2 to python3.

I have verified that julia-1.0 builds with this change.  julia-1.5 is marked as broken
in nixpkgs, so I have not tested it (but I would be very surprised if this change
broke it further).

Motivated by #148779