gitin: remove at 0.2.3
gitin depends on a precise version of git2go. Not even the latest
gitin/git2go version is compatible with an up to date libgit2. As the
upstream project hasn't seen activity in some time and doesn't respond
to the related issue (https://github.com/isacikgoz/gitin/issues/71),
drop this package.