dpdk: add Mellanox support via rdma-core dependency
rdma-core contains libmlx4/libmlx5 which are required for DPDK to be
able to run using Mellanox NIC devices.

This has a non-trivial closure size impact (730M -> 941M) but the size
increase is almost entirely driven by adding systemd to the transitive
closure, which we should expect is present anyway on most NixOS systems.
So, in practice, this is expected to be close to no-op in size increase.