nixos: force the user to set system.stateVersion
The argument for introducing `system.stateVersion` was that some data
that lives on disk, like the postgres database, would change shape
depending on which release of NixOS the user was using.

If the default is always the current version, it can lead to unwanted
breakages. It's better to fail the evaluation and force the user to set
something.