nixops: remove (v1.x)
The NixOps 1.x branch has been unsupported for a long time, breaks
very often in Nixpkgs, because it requires heavily outdated deps,
and has put undue burden on the python packages maintainers.

The 1.x branch is unsustainable and it is time that we acknowledge
this fact. The only way forward for NixOps is the unreleased
version 2.0, which needs testing and fixes before it can be
released and become the replacement for the default nixops package.